PARTY/backend/requirements.txt
black==22.3.0
blessings==1.7
bpython==0.22.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
curtsies==0.3.10
cwcwidth==0.1.6
darker==1.4.2
Django>=4.0.7,<4.1
django-cors-headers==3.11.0
django-dynamic-filenames==1.3.0
django-extensions==3.1.5
django-filter==21.1
django-leaflet==0.28.2
django-stdimage==5.3.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
drf-spectacular==0.22.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
isort==5.10.1
jedi==0.18.1
jsonschema==4.6.1
mypy-extensions==0.4.3
parso==0.8.3
pathspec==0.9.0
Pillow==9.2.0
platformdirs==2.5.2
pluggy==1.0.0
psycopg2-binary==2.9.3
Pygments==2.11.2
PyJWT==2.3.0
pyrsistent==0.18.1
python-decouple==3.6
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.4.1
pytz==2022.1
pyxdg==0.27
PyYAML==6.0
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
typing_extensions==4.2.0
ujson==5.2.0
uritemplate==4.1.1
urllib3==1.26.9
Werkzeug==2.1.1
django-rest-passwordreset==1.3.0
six==1.16.0